requirements.txt
streamlit
graphviz
google-generativeai
python-dotenv
pathlib
seaborn
streamlit_lottie
-e .